CARL ResourceManager: Move `subscription` module into `api`.
