+++ opendut-carl/src/resource/api/resources.rs
@@ -2,7 +2,7 @@ use crate::resource::api::Resource;
 use crate::resource::persistence::error::PersistenceResult;
 use crate::resource::persistence::persistable::Persistable;
 use crate::resource::storage::{ResourceTransaction, ResourcesStorageApi};
-use crate::resource::subscription::{ResourceSubscriptionChannels, Subscribable, SubscriptionEvent};
+use crate::resource::api::subscription::{ResourceSubscriptionChannels, Subscribable, SubscriptionEvent};
 use std::collections::HashMap;
 
 pub type RelayedSubscriptionEvents = ResourceSubscriptionChannels;
